% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derive_ExtremeValueFlag.R
\name{DefaultExtremeParameters}
\alias{DefaultExtremeParameters}
\title{The default map from Table 58 to 60 parameters to \code{ExampleData("adbds")} test names}
\usage{
DefaultExtremeParameters()
}
\value{
Named \code{list} of \code{character}.
}
\description{
The default map from Table 58 to 60 parameters to \code{ExampleData("adbds")} test names
}
\keyword{internal}
