% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_SafetyCensus.R
\name{.CensusFirstValue}
\alias{.CensusFirstValue}
\title{The first non-missing value of the first column present}
\usage{
.CensusFirstValue(dfResults, chrCols)
}
\description{
The first non-missing value of the first column present
}
\keyword{internal}
