% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report_SafetyCensus.R
\name{.CensusStudyID}
\alias{.CensusStudyID}
\title{The study identifier the results carry}
\usage{
.CensusStudyID(dfResults)
}
\description{
The study identifier the results carry
}
\keyword{internal}
