% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafetyCensus.R
\name{.LegacyCensus}
\alias{.LegacyCensus}
\title{The \code{Census} table}
\usage{
.LegacyCensus(dfFigures)
}
\arguments{
\item{dfFigures}{\code{data.frame} From \code{\link[=Report_CensusFigures]{Report_CensusFigures()}}.}
}
\value{
\code{data.frame} with \code{Label}, \code{Value}, \code{Denominator} and \code{Group}.
}
\description{
The \code{Census} table
}
\keyword{internal}
