% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafetyCensus.R
\name{.LegacyDisposition}
\alias{.LegacyDisposition}
\title{The \code{Disposition} table}
\usage{
.LegacyDisposition(dfFigures)
}
\arguments{
\item{dfFigures}{\code{data.frame} From \code{\link[=Report_CensusFigures]{Report_CensusFigures()}}.}
}
\value{
\code{data.frame} with \code{State} and \code{Participants}, largest first.
}
\description{
The \code{Disposition} table
}
\keyword{internal}
