% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafetyCensus.R
\name{.VisitCoverage}
\alias{.VisitCoverage}
\title{Participants with at least one record, per visit}
\usage{
.VisitCoverage(df, strDomain, strIDCol, strVisitCol, strVisitNumCol, nExpected)
}
\description{
Participants with at least one record, per visit
}
\keyword{internal}
