% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derive_AbnormalityLevel.R
\name{NormaliseSex}
\alias{NormaliseSex}
\title{Fold the spellings of sex to \code{m} or \code{f}}
\usage{
NormaliseSex(chrSex)
}
\arguments{
\item{chrSex}{\code{character} Sex values.}
}
\value{
\code{character} \code{"m"}, \code{"f"} or \code{NA}.
}
\description{
Fold the spellings of sex to \code{m} or \code{f}
}
\keyword{internal}
