% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-derive.R
\name{RequireResultColumns}
\alias{RequireResultColumns}
\title{Require a data frame with the named columns}
\usage{
RequireResultColumns(df, chrCols, strName = "dfResults")
}
\arguments{
\item{df}{The object to check.}

\item{chrCols}{\code{character} Column names that must be present.}

\item{strName}{\code{character} Name of the argument, for the message.}
}
\description{
Require a data frame with the named columns
}
\keyword{internal}
