% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget_ParticipantProfile.R
\name{Widget_ParticipantProfile}
\alias{Widget_ParticipantProfile}
\title{Participant Profile Widget (PROBE: naive, single-dfResults contract)}
\usage{
Widget_ParticipantProfile(
  dfResults,
  lSettings = list(),
  width = NULL,
  height = NULL,
  elementId = NULL,
  bDebug = FALSE
)
}
\description{
Participant Profile Widget (PROBE: naive, single-dfResults contract)
}
